requirements.txt
# Mantissa Log - Production Dependencies
# Python 3.11+

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data processing
pyyaml>=6.0.1
jsonschema>=4.20.0

# HTTP and API
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.2

# Logging and monitoring
structlog>=24.1.0

# LLM providers
anthropic>=0.8.0
openai>=1.0.0

# Parser dependencies
ipaddress>=1.0.23
user-agents>=2.2.0
